Major improvements

+++ src/manager/src/modules/server/socket/development.js
@@ -1,34 +1,74 @@
 var async = require('async'),
-    gulp = require('bluestorm').gulp,
-    childProcess = require('child_process');
+gulp = require('bluestorm').gulp,
+childProcess = require('child_process'),
+config = require(__dirname+'/../../config/lib/config');
 
-var child;
+var child,
+apps;
 
 module.exports = function(socket) {
     socket.on('server:development:start', function(req, callback) {
         if (typeof callback == 'function') {
             process.env.NODE_ENV = 'development';
 
-            /*if(child) {
+            if(child) {
                 child.kill('SIGHUP');
             }
-            child = childProcess.fork(__dirname+'/../../../../lib/gulp/development', {
+            child = childProcess.fork(__dirname+'/../lib/development', {
                 cwd: req.data.path
             });
             child.on('message', function(message){
-                socket.emit('message', message);
+                if(message.type=='app_started') {
+                    apps.forEach(function (app) {
+                        if(app.name==message.name) {
+                            app.status = 'up'
+                        };
+                    });
+                }
+                socket.emit('message_server', message);
             });
             child.send({
-                debug: true
-            });*/
+                debug: true,
+                apps: req.data.apps
+            });
 
             callback();
         }
     });
-    /*
 
-    
-    socket.on('tasks:kill', function(req, callback) {
+    socket.on('server:getApps', function(req, callback) {
+        if (typeof callback == 'function') {
+            if(apps) {
+                callback(null, apps);
+            } else {
+
+                apps = [{
+                    name: 'api',
+                    status: 'down'
+                }, {
+                    name: 'socket',
+                    status: 'down'
+                }];
+                config(req.data.path).frontend.list().forEach(function (app) {
+                    apps.push({
+                        name: app,
+                        status: 'down'
+                    })
+                });
+                callback(null, apps);
+            }
+        }
+    });
+    socket.on('server:isProcessing', function(req, callback) {
+        if (typeof callback == 'function') {
+            if(child) {
+                callback(null, true)
+            } else {
+                callback(null, false)
+            }
+        }
+    });
+    socket.on('server:kill', function(req, callback) {
         if (typeof callback == 'function') {
             if(child) {
                 child.kill('SIGHUP');
@@ -40,13 +80,4 @@ module.exports = function(socket) {
             callback();
         }
     });
-    socket.on('tasks:isProcessing', function(req, callback) {
-        if (typeof callback == 'function') {
-            if(child) {
-                callback(null, true)
-            } else {
-                callback(null, false)
-            }
-        }
-    });*/
 };
